restructure app routes to support account pages
