+++ src/app/(accounts)/accounts/layout.tsx
@@ -1,0 +1,8 @@
+
+export default function RootLayout({
+    children,
+}: {
+  children: React.ReactNode
+}) {
+    return children
+}
