+++ src/app/(accounts)/accounts/login/page.tsx
@@ -1,0 +1,9 @@
+
+
+const LoginPage = () => {
+    return (
+        <div>sag</div>
+    )
+}
+
+export default LoginPage
