+++ src/app/(app)/layout.tsx
@@ -1,6 +1,6 @@
 import type {Metadata} from 'next'
 import {Manrope} from 'next/font/google'
-import './globals.css'
+import '../globals.css'
 import Navbar from "@/components/Navbar"
 import Footer from "@/components/Footer"
 import {Separator} from "@/components/ui/separator"
