+++ src/app/(app)/project-sets/page.tsx
@@ -1,10 +1,10 @@
 import React from "react"
-import {columns} from "@/app/project-sets/columns"
 import {DataTable} from "@/components/ui/data-table"
 import PageView from "@/components/views/Page"
 import {type ApiTeamSetTemplate} from "@/_temp_types/api/teams"
 import {type ProjectSet} from "@/_temp_types/projectSet"
 import {redirect} from "next/navigation"
+import { columns } from "./columns"
 
 async function getProjectSetsData(): Promise<ProjectSet[]> {
     const response = await fetch(process.env.BACKEND_URL + '/api/v1/teamset-templates')
